Add files via upload
Hello! Every time I run it with the bonus code it goes to "not responding" on both excel and VBA. I have my code copied onto a word document every time in order to avoid the risk of it not saving and will post the excel screenshots, but I am not sure the VBA script will run with the bonus because I have not been able to successfully run the code. I will upload it anyways. Wanted to place this in comment in case there was confusion as to why I uploaded a word document. Thanks and have a nice weekend.